default value in select-field
When select-field is used with undefined/null value, the first option is shown. But the value is not set to this value unless we change the options.

A solution is to work with '' as with default value in route-action but then we can run into inconsistencies. Additional argument in template should fix that

https://github.com/marxsk/picus/issues/41